+++ .gitignore
@@ -4,6 +4,8 @@ config.py
 __pycache__/
 *.py[cod]
 *$py.class
+service_account_creds.json
+other/
 
 # C extensions
 *.so
