+++ placeholder.env
@@ -1,0 +1,3 @@
+PROJECTID = 'enter gcp PROJECTID'
+DATASET = 'enter gcp DATASET'
+TABLENAME = 'enter gcp TABLENAME'
